backend/requirements.txt
# Core backend
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1

# OpenAI client
openai==1.107.1
